% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{biplot_colored}
\alias{biplot_colored}
\title{Another implementation of the biplot function}
\usage{
biplot_colored(x, y, rank = TRUE, ties_method = c("max", "min", "first",
  "last", "random"), choices = 1:2, expand = 1, ...)
}
\arguments{
\item{x}{the result of a call to \code{\link[stats]{prcomp}}.}

\item{y}{an array of values used to color the points. 
If rank is FALSE, all values must be positive integers and less than or equal to the length of y.}

\item{rank}{logical. If TRUE (default) y will be transformed by the rank() function}

\item{ties_method}{character. ties.method used by the rank() function}

\item{choices}{which principal components to plot. Only 2D plots are
possible for now. Default to first two PCs.}

\item{expand}{numeric value used to adjust the spread of the arrows relative
to the points.}

\item{...}{passed to plot.}
}
\description{
Function to plot a biplot with no point labels and with points color-coded
according to a certain quantitative variable, for instance the rank of the
normalization performance or the expression of a certain gene.
}
\details{
This function implements the biplot only for \code{\link[stats]{prcomp}}
objects. Eventually, we will turn this into an S4 method.
}
\examples{
mat <- matrix(rnorm(1000), ncol=10)
colnames(mat) <- paste("X", 1:ncol(mat), sep="")

pc <- prcomp(mat)

biplot_colored(pc, rank(pc$x[,1]))

}

