% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot_interactive.R
\name{biplot_interactive}
\alias{biplot_interactive}
\title{Interactive biplot}
\usage{
biplot_interactive(data, scores, ...)
}
\arguments{
\item{data}{a data.frame containing the data to be plotted.}

\item{scores}{a numeric vector used to color the points.}
}
\description{
This is a wrapper around \code{\link{biplot_colored}}, which creates a shiny
gadget to allow the user to select specific points in the graph.
}
\details{
Since this is based on the shiny gadget feature, it will not work in
  static documents, such as vignettes or markdown / knitr documents.
  See \code{biplot_colored} for more details on the internals.
}
\examples{
\dontrun{
mat <- matrix(rnorm(1000), ncol=10)
colnames(mat) <- paste("X", 1:ncol(mat), sep="")

biplot_interactive(mat, mat[,1])
}
}

