% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCONE_DEFAULTS.R
\name{DESEQ_FN}
\alias{DESEQ_FN}
\title{DESeq normalization wrapper.}
\usage{
DESEQ_FN(ei)
}
\arguments{
\item{ei}{= Numerical matrix. (rows = genes, cols = samples). Unique row.names are required.}
}
\value{
DESeq normalized matrix (scaled by sample )
}
\description{
DESeq normalization wrapper.
}

