% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCONE_DEFAULTS.R
\name{DESEQ_FN_POS}
\alias{DESEQ_FN_POS}
\title{DESeq size factor normalization derived from positive data.}
\usage{
DESEQ_FN_POS(ei)
}
\arguments{
\item{ei}{Numerical matrix. (rows = genes, cols = samples).}
}
\value{
DESeq size factor (positive) normalized matrix.
}
\description{
DESeq size factor normalization derived from positive data.
}
\details{
SCONE scaling function scales expression data by DESeq size factor derived from positive data.
}

