% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.parse_row}
\alias{.parse_row}
\title{Parse rows}
\usage{
.parse_row(pars, bio, batch, ruv_factors, qc)
}
\arguments{
\item{pars}{character. A vector of parameters corresponding to a row of
workflow parameters.}

\item{bio}{factor. The biological covariate.}

\item{batch}{factor. The batch covariate.}

\item{ruv_factors}{list. A list containing the factors of unwanted variation
(RUVg) for all upstream workflows.}

\item{qc}{matrix. The principal components of the QC metric matrix.}
}
\value{
A list with the variables to be passed to make_design.
}
\description{
This function is used internally in scone to parse the variables used to
generate the design matrices.
}
\keyword{internal}
