% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ziber.R
\name{.pzfn}
\alias{.pzfn}
\title{Posterior probability of detection}
\usage{
.pzfn(Y, W, Alpha, X, Beta)
}
\arguments{
\item{Y}{detection matrix.}

\item{W}{sample-level drop-out coefficients.}

\item{Alpha}{gene-level drop-out features.}

\item{X}{sample-level expression features.}

\item{Beta}{gene-level sample coefficients.}
}
\description{
Posterior probability of detection
}
\keyword{internal}
