% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb.R
\name{estimate_zinb}
\alias{estimate_zinb}
\title{Parameter estimation of zero-inflated negative binomial model}
\usage{
estimate_zinb(Y, maxiter = 10, verbose = FALSE)
}
\arguments{
\item{Y}{the data matrix (genes in rows, cells in columns)}
}
\value{
a list with the following elements:
\itemize{
\item{mu}{the mean of the negative binomial component}
\item{theta}{the dispersion parameter of the negative binomial component}
\item{pi}{the mixing probability}
\item{coefs}{the coefficients of the logistic regression}
\item{p_z}{the probability P(Z=0|Y), useful as weights in weighted principal components analysis}
\item{expected_value}{the expected value E[Y]}
\item{variance}{the variance Var(Y)}
\item{loglik}{the log-likelihood}
\item{convergence}{0 if the algorithm converged and 1 if maxiter was reached}
}
}
\description{
This function implements an EM algorithm to estimate the parameters of a zero-inflated negative binomial model
}
\details{
This function implements an expectation-maximization algorithm for a zero-inflated
model, with a constant gene-specific mean parameter, gene-specific dispersion and
a mixing probability that depends solely on the mean.
}

