% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCONE_DEFAULTS.R
\name{FQ_FN_POS}
\alias{FQ_FN_POS}
\title{Full-Quantile normalization applied to positive data.}
\usage{
FQ_FN_POS(ei)
}
\arguments{
\item{ei}{= Numerical matrix. (rows = genes, cols = samples). Unique row.names are required.}
}
\value{
FQ (positive) normalized matrix.
}
\description{
Full-Quantile normalization applied to positive data.
}

