% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper.R
\docType{methods}
\name{get_bio}
\alias{get_batch}
\alias{get_batch,SconeExperiment-method}
\alias{get_bio}
\alias{get_bio,SconeExperiment-method}
\title{Get Factor of Biological Conditions and Batch}
\usage{
get_bio(x)

get_batch(x)

\S4method{get_bio}{SconeExperiment}(x)

\S4method{get_batch}{SconeExperiment}(x)
}
\arguments{
\item{x}{an object of class \code{\link{sconeExperiment}}.}
}
\value{
NULL or a factor containing bio or batch covariate.
}
\description{
Get Factor of Biological Conditions and Batch
}

