% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper.R
\docType{methods}
\name{get_negconruv}
\alias{get_negconeval}
\alias{get_negconeval,SconeExperiment-method}
\alias{get_negconruv}
\alias{get_negconruv,SconeExperiment-method}
\alias{get_poscon}
\alias{get_poscon,SconeExperiment-method}
\title{Get Negative and Positive Controls}
\usage{
get_negconruv(x)

get_negconeval(x)

get_poscon(x)

\S4method{get_negconruv}{SconeExperiment}(x)

\S4method{get_negconeval}{SconeExperiment}(x)

\S4method{get_poscon}{SconeExperiment}(x)
}
\arguments{
\item{x}{an object of class \code{\link{sconeExperiment}}.}
}
\value{
NULL or a logical vector.

For \code{get_negconruv} it indicates which genes are negative controls
  to be used for RUV.

For \code{get_negconeval} it indicates which genes are negative controls
  to be used for evaluation.

For \code{get_poscon} it indicates which genes are positive controls.
}
\description{
Get Negative and Positive Controls
}

