% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scone_main.R
\name{get_normalized}
\alias{get_normalized}
\title{Retrieve Normalized Matrix}
\usage{
get_normalized(normalization, hdf5file)
}
\arguments{
\item{normalization}{character. A string identifying the normalization scheme
to be retrieved.}

\item{hdf5file}{character. The name of the file in which the data have been
stored.}
}
\value{
A matrix of normalized counts in log-scale.
}
\description{
Given a string representing a normalization scheme and an HDF5 file created by
scone, it will return a matrix of normalized counts (in log scale).
}
\details{
The string must be of of the row.names of the \code{params}
  component of the scone output.
}

