% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper.R
\docType{methods}
\name{get_params}
\alias{get_params}
\alias{get_params,SconeExperiment-method}
\title{Extract scone parameters}
\usage{
get_params(x)

\S4method{get_params}{SconeExperiment}(x)
}
\arguments{
\item{x}{an object of class \code{\link{sconeExperiment}}.}
}
\value{
A data.frame with the information of the normalization schemes that
  scone has run.
}
\description{
Extract scone parameters
}

