% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper.R
\docType{methods}
\name{get_qc}
\alias{get_qc}
\alias{get_qc,SconeExperiment-method}
\title{Get Quality Control Matrix}
\usage{
get_qc(x)

\S4method{get_qc}{SconeExperiment}(x)
}
\arguments{
\item{x}{an object of class \code{\link{sconeExperiment}}.}
}
\value{
NULL or the QC matrix.
}
\description{
Get Quality Control Matrix
}

