% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper.R
\docType{methods}
\name{get_scores}
\alias{get_score,SconeExperiment-method}
\alias{get_score_ranks}
\alias{get_score_ranks,SconeExperiment-method}
\alias{get_scores}
\alias{get_scores,SconeExperiment-method}
\title{Extract scone scores}
\usage{
get_scores(x)

get_score_ranks(x)

\S4method{get_scores}{SconeExperiment}(x)

\S4method{get_score_ranks}{SconeExperiment}(x)
}
\arguments{
\item{x}{an object of class \code{\link{sconeExperiment}}.}
}
\value{
\code{get_scores} returns a matrix with the non-missing scone scores,
  ordered by average score.

\code{get_score_ranks} returns a vector with the ranks of the average
  score.
}
\description{
Extract scone scores
}

