% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ziber.R
\name{impute_null}
\alias{impute_null}
\title{Null or no-op imputation}
\usage{
impute_null(expression, impute_args)
}
\arguments{
\item{expression}{the data matrix (genes in rows, cells in columns)}

\item{impute_args}{arguments for imputation (not used)}
}
\value{
the imputed expression matrix.
}
\description{
Null or no-op imputation
}
\examples{
mat <- matrix(rpois(1000, lambda = 5), ncol=10)
imat = impute_null(mat)

}

