% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ziber.R
\name{impute_ziber_simp}
\alias{impute_ziber_simp}
\title{Imputation of zero abundance based on zero-inflated bernoulli model}
\usage{
impute_ziber_simp(expression)
}
\arguments{
\item{expression}{the data matrix (genes in rows, cells in columns)}
}
\value{
the imputed expression matrix.
}
\description{
This function is used to impute the data, by weighting the observed detection failure by their
probability of coming from the zero-inflation part of the distribution. Median positive expression
is used as the sole gene-level determinant of drop-out, and the data is replaced with this expression value.
}
\details{
The imputation is carried out with the following formula:
y_{ij}* =  y_{ij} * Pr( No Drop | D_{ij}) + pos_median_{i} * Pr( Drop | D_{ij}).
}

