% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinb.R
\name{impute_zinb}
\alias{impute_zinb}
\title{Imputation of zero counts based on zero-inflated negative binomial model}
\usage{
impute_zinb(expression)
}
\arguments{
\item{expression}{the data matrix (genes in rows, cells in columns)}
}
\value{
the imputed expression matrix.
}
\description{
This function is used to impute the data, by weighting the observed zeros by their
probability of coming from the zero-inflation part of the distribution P(Z=1|Y).
}
\details{
The imputation is carried out with the following formula:
y_{ij}* =  y_{ij} * Pr(Z_{ij} = 0 | Y_{ij}=y_{ij}) + mu_{ij} * Pr(Z_{ij} = 1 | Y_{ij} = y_{ij}).
Note that for y_{ij} > 0, Pr(Z_{ij} = 0 | Y_{ij}=y_{ij}) = 1 and hence the data are not imputed.
}

