% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ziber.R
\name{likfn}
\alias{likfn}
\title{Likelihood Function of the Logistic Model}
\usage{
likfn(Z, X, Beta)
}
\arguments{
\item{Z}{data matrix}

\item{X}{sample-level values}

\item{Beta}{gene-level values}
}
\description{
Likelihood Function of the Logistic Model
}

