% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{lm_adjust}
\alias{lm_adjust}
\title{Function to perform linear batch effect correction}
\usage{
lm_adjust(log_expr, design_mat, batch = NULL, weights = NULL)
}
\arguments{
\item{log_expr}{matrix. The log gene expression (genes in row, samples in columns).}

\item{design_mat}{matrix. The design matrix (usually the result of make_design).}

\item{batch}{factor. A factor with the batch information.}

\item{weights}{matrix. A matrix of weights.}
}
\value{
The corrected log gene expression.
}
\description{
Given a matrix with log expression values and a design matrix, this function fits a linear model
and removes the effects of the batch factor as well as of the linear variables encoded in W.
}
\details{
The function assumes that the columns of the design matrix corresponding to the variable
for which expression needs to be adjusted, start with either the word "batch" or the letter "W" (case sensitive).
Any other covariate (including the intercept) is kept.
}

