% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCONE_DEFAULTS.R
\name{LSF_FN}
\alias{LSF_FN}
\title{LSF normalization wrapper.}
\usage{
LSF_FN(ei)
}
\arguments{
\item{ei}{= Numerical matrix. (rows = genes, cols = samples). Unique row.names are required.}
}
\value{
LSF normalized matrix (scaled by sample )
}
\description{
SCONE scaling wrapper for \code{\link[scran]{computeSumFactors}}
  with clusters from \code{\link[scran]{quickCluster}}.
}

