% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_filtering.R
\name{metric_sample_filter}
\alias{metric_sample_filter}
\title{metric-based sample filtering: function to filter single-cell RNA-Seq libraries.}
\usage{
metric_sample_filter(expr, nreads = NULL, ralign = NULL,
  gene_filter = NULL, pos_controls = NULL, glen = NULL, AUC_range = c(0,
  6), zcut = 1, mixture = TRUE, dip_thresh = 0.05, hard_nreads = 25000,
  hard_ralign = 15, hard_breadth = 0.2, hard_fnr = 3,
  suff_nreads = NULL, suff_ralign = 65, suff_breadth = 0.8,
  suff_fnr = NULL, plot_dir = NULL, hist_breaks = 10)
}
\arguments{
\item{expr}{matrix The data matrix (genes in rows, cells in columns).}

\item{nreads}{A numeric vector representing number of reads in each library.
If NULL, filtered_nreads will be returned NA.}

\item{ralign}{A numeric vector representing the proportion of reads aligned to the reference genome in each library.
If NULL, filtered_ralign will be returned NA.}

\item{gene_filter}{A boolean vector indexing genes that will be used to compute library transcriptome breadth.
If NULL, filtered_breadth will be returned NA.}

\item{pos_controls}{A boolean vector indexing positive control genes that will be used to compute false-negative rate characteristics.
If NULL, filtered_fnr will be returned NA.}

\item{glen}{Gene lengths for gene-length normalization (normalized data used in FNR computation).}

\item{AUC_range}{An array of two values, representing range over which FNR AUC will be computed (log10(expr_units + 1)). Default c(0,6)}

\item{zcut}{A numeric value determining threshold Z-score for sd, mad, and mixture sub-criteria. Default 1.
If NULL, only hard threshold sub-criteria will be applied.}

\item{mixture}{A boolean value determining whether mixture modeling sub-criterion will be applied per primary criterion (metric).
If true, a dip test will be applied to each metric. If a metric is multimodal, it is fit to a two-component nomal mixture model. 
Samples deviating zcut sd's from optimal mean (in the inferior direction), have failed this sub-criterion.}

\item{dip_thresh}{A numeric value determining dip test p-value threshold. Default 0.05.}

\item{hard_nreads}{numeric. Hard (lower bound on) nreads threshold. Default 25000.}

\item{hard_ralign}{numeric. Hard (lower bound on) ralign threshold. Default 15.}

\item{hard_breadth}{numeric. Hard (lower bound on) breadth threshold. Default 0.2.}

\item{hard_fnr}{numeric. Hard (upper bound on) fnr threshold. Default 3.}

\item{suff_nreads}{numeric. If not null, serves as an upper bound on nreads threshold.}

\item{suff_ralign}{numeric. If not null, serves as an upper bound on ralign threshold.  Default 65.}

\item{suff_breadth}{numeric. If not null, serves as an upper bound on breadth threshold.  Default 0.8.}

\item{suff_fnr}{numeric. If not null, serves as an lower bound on fnr threshold.}

\item{plot_dir}{If not null, specifies path to plot output}

\item{hist_breaks}{hist() breaks argument}
}
\value{
A list with the following elements:
\itemize{
\item{filtered_nreads}{Boolean. Sample has too few reads.}
\item{filtered_ralign}{Boolean. Sample has too few reads aligned.}
\item{filtered_breadth}{Boolean. Samples has too few genes detected (low breadth).}
\item{filtered_fnr}{Boolean. Sample has a high FNR AUC.}
}
}
\description{
This function returns a sample-filtering report for each cell in the input expression matrix, 
describing which filtering criteria are satisfied.
}
\details{
For each primary criterion (metric), a sample is evaluated based on 4 sub-criteria: 
1) Hard (encoded) threshold 
2) Adaptive thresholding via sd's from the mean
3) Adaptive thresholding via mad's from the median
4) Adaptive thresholding via sd's from the mean (after mixture modeling)
A sample must pass all sub-criteria to pass the primary criterion.
}

