% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{parse_row}
\alias{parse_row}
\title{Internal Function}
\usage{
parse_row(pars, bio, batch, ruv_factors, qc)
}
\arguments{
\item{pars}{character. A vector of parameters corresponding to a row of params.}

\item{bio}{factor. The biological factor of interest.}

\item{batch}{factor. The known batch effects.}

\item{ruv_factors}{list. A list containing the factors of unwanted variation.}

\item{qc}{matrix. The principal components of the QC metrics.}
}
\value{
A list with the variables to be passed to make_design.
}
\description{
This function is used internally in scone to parse the variables used to generate the design matrices.
}

