% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCONE_DEFAULTS.R
\name{PSINORM_FN}
\alias{PSINORM_FN}
\title{PsiNorm normalization wrapper}
\usage{
PSINORM_FN(ei)
}
\arguments{
\item{ei}{Numerical matrix. (rows = genes, cols = samples).}
}
\value{
PsiNorm normalized matrix.
}
\description{
PsiNorm normalization wrapper
}
\details{
SCONE scaling wrapper for \code{\link{PsiNorm}}).
}
\examples{
ei <- matrix(c(1,0,2,0,2,9,3,0), ncol=2)
eo <- PSINORM_FN(ei)

}
