% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scone_main.R
\name{scone}
\alias{scone}
\title{scone main wrapper: function to apply and evaluate all the normalization schemes}
\usage{
scone(expr, imputation, scaling, k_ruv = 5, k_qc = 5, ruv_negcon = NULL,
  qc = NULL, adjust_bio = c("no", "yes", "force"), adjust_batch = c("no",
  "yes", "force"), bio = NULL, batch = NULL, evaluate = TRUE,
  eval_pcs = 3, eval_knn = 10, eval_kclust = 2:10, eval_negcon = NULL,
  eval_poscon = NULL, run = TRUE, params = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{matrix. The data matrix (genes in rows, cells in columns).}

\item{imputation}{list or function. (A list of) function(s) to be used for imputation.}

\item{scaling}{list or function. (A list of) function(s) to be used for scaling normalization.}

\item{k_ruv}{numeric. The maximum number of factors of unwanted variation (the function will adjust for 1 to k_ruv factors of unwanted variation).
If 0, RUV will not be performed.}

\item{k_qc}{numeric. The maximum number of quality metrics PCs (the function will adjust for 1 to k_qc PCs).
If 0, QC adjustment will not be performed.}

\item{ruv_negcon}{character. The genes to be used as negative controls for RUV. Ignored if k_ruv=0.}

\item{qc}{matrix. The QC metrics to be used for QC adjustment. Ignored if k_qc=0.}

\item{adjust_bio}{character. If 'no' it will not be included in the model; if 'yes', both models with and without 'bio' will be run;
if 'force', only models with 'bio' will be run.}

\item{adjust_batch}{character. If 'no' it will not be included in the model; if 'yes', both models with and without 'batch' will be run;
if 'force', only models with 'batch' will be run.}

\item{bio}{factor. The biological condition to be included in the adjustment model (variation to be preserved).
Ignored, if adjust_bio=0.}

\item{batch}{factor. The known batch variable to be included in the adjustment model (variation to be removed).
Ignored, if adjust_batch=0.}

\item{evaluate}{logical. If FALSE the normalization methods will be run but not evaluated.}

\item{eval_pcs}{numeric. The number of principal components to use for evaluation. Ignored if evaluation=FALSE.}

\item{eval_knn}{numeric. The number of nearest neighbors to use for evaluation. Ignored if evaluation=FALSE.}

\item{eval_kclust}{numeric. The number of clusters (> 1) to be used for pam stability evaluation. If NULL, all KNN concordances will be returned NA.
If an array of integers, largest average silhoutte width will be reported. If NULL, stability will be returned NA.}

\item{eval_negcon}{character. The genes to be used as negative controls for evaluation. These genes should
be expected not to change according to the biological phenomenon of interest. Ignored if evaluation=FALSE.
If NULL, correlations with negative controls will be returned NA.}

\item{eval_poscon}{character. The genes to be used as positive controls for evaluation. These genes should
be expected to change according to the biological phenomenon of interest. Ignored if evaluation=FALSE.
If NULL, correlations with positive controls will be returned NA.}

\item{run}{logical. If FALSE the normalization and evaluation are not run, but the function returns a matrix
of parameters that will be run for inspection by the user.}

\item{params}{matrix or data.frame. If given, the algorithm will bypass creating the matrix of possible
parameters, and will use the given matrix. There are basically no checks as to whether this matrix is in the
right format, and is only intended to be used to feed the results of setting run=FALSE back into 
the algorithm (see example).}

\item{verbose}{logical. If TRUE some messagges are printed.}
}
\value{
If run=TRUE, a list with the following elements:
\itemize{
\item{normalized_data}{a list containing the normalized data (after scaling and factor adjustment).}
\item{design_matrix}{a list with the design matrix used for the factor adjustments.}
\item{ruv_factors}{a list with the RUV factors computed from each combination of imputation and scaling.}
}
If run=FALSE, a data.frame with each row corresponding to a set of parameters to be analyzed.
}
\description{
This function is a high-level wrapper to apply and evaluate a variety of normalization
schemes to a specified expression matrix.
}
\details{
The function consists of four main steps: imputation, scaling normalization, adjusting for batch effects,
and evaluation of the normalization performance.

Note that if one wants to include the unnormalized data in the comparison, the identity function
must be included in the scaling list. Analogously, if one wants to avoid the imputation and/or include the
non-imputed data in the comparison, the identity function must be included.
}

