% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SconeExperiment-class}
\alias{SconeExperiment}
\alias{SconeExperiment-class}
\title{Class SconeExperiment}
\description{
Objects of this class store,  at the minimum, the original gene
  expression matrix and a set of covariates (aka sample metadata) useful for
  running \code{\link{scone}}. These include, the quality control (QC)
  metrics, batch information, and biological classes of interest (if
  available).

The typical way of creating \code{SconeExperiment} objects is
  via a call to the \code{\link{sconeExperiment}} function or to the
  \code{\link{scone}} function. If the object is a result to a
  \code{\link{scone}} call, it will contain the results, e.g., the
  performance metrics, scores, and the normalization schemes compared. (See
  Slots for a full list).

This object extends the
  \code{\linkS4class{SummarizedExperiment}} class.
}
\details{
The QC matrix, biological class, and batch information are stored as
  elements of the `colData` of the object.

The positive and negative control genes are stored as elements of
  the `rowData` of the object.
}
\section{Slots}{

\describe{
\item{\code{which_qc}}{integer. Index of columns of `colData` that contain the QC
metrics.}

\item{\code{which_bio}}{integer. Index of the column of `colData` that contains the
biological classes information (it must be a factor).}

\item{\code{which_batch}}{integer. Index of the column of `colData` that contains the
batch information (it must be a factor).}

\item{\code{which_negconruv}}{integer. Index of the column of `rowData` that contains
a logical vector indicating which genes to use as negative controls to
infer the factors of unwanted variation in RUV.}

\item{\code{which_negconeval}}{integer. Index of the column of `rowData` that
contains a logical vector indicating which genes to use as negative
controls to evaluate the performance of the normalizations.}

\item{\code{which_poscon}}{integer. Index of the column of `rowData` that contains a
logical vector indicating which genes to use as positive controls to
evaluate the performance of the normalizations.}

\item{\code{hdf5_pointer}}{character. A string specifying to which file to write /
read the normalized data.}

\item{\code{scone_metrics}}{matrix. Matrix containing the "raw" performance metrics.
See \code{\link{scone}} for a description of each metric.}

\item{\code{scone_scores}}{matrix. Matrix containing the performance scores
(transformed metrics). See \code{\link{scone}} for a discussion on the
difference between scores and metrics.}

\item{\code{scone_parmas}}{data.frame. A data frame containing the normalization
schemes applied to the data and compared.}

\item{\code{design_mats}}{list. A list of design matrices, one for each normalization
scheme.}

\item{\code{scone_run}}{character Whether \code{\link{scone}} and in which mode
("no", "in_memory", "hdf5").}
}}

