% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCONE_DEFAULTS.R
\name{SCRAN_FN}
\alias{SCRAN_FN}
\title{Pooled Sample normalization wrapper.}
\usage{
SCRAN_FN(ei)
}
\arguments{
\item{ei}{Numerical matrix. (rows = genes, cols = samples).}
}
\value{
SCRAN size factor normalized matrix.
}
\description{
SCONE scaling wrapper for \code{\link[scran]{computeSumFactors}}
  with clusters from \code{\link[scran]{quickCluster}}.
}

