% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sconeReport.R
\name{subsample_cells_with_min_bio}
\alias{subsample_cells_with_min_bio}
\title{Internal function to subsample Scone object by subsampling cells by size of smallest bio group}
\usage{
subsample_cells_with_min_bio(scone_object, seed = 100, verbose = FALSE)
}
\arguments{
\item{scone_object}{a \code{SconeExperiment} object}

\item{seed}{the random seed}

\item{verbose}{option to display more information about subsampling}
}
\value{
The subsampled Scone Object
}
\description{
This function subsamples a scone object to the size of smallest bio group
}
\examples{
# Create a scone object
mat <- matrix(rpois(1000, lambda = 5), ncol=10)
colnames(mat) <- paste("X", 1:ncol(mat), sep="")
obj <- SconeExperiment(mat)
res <- scone(obj, scaling=list(none=identity, uq=UQ_FN, deseq=DESEQ_FN),
           evaluate=TRUE, k_ruv=0, k_qc=0, eval_kclust=2,
           bpparam = BiocParallel::SerialParam())
qc = as.matrix(cbind(colSums(mat),colSums(mat > 0)))
rownames(qc) = colnames(mat)
colnames(qc) = c("NCOUNTS","NGENES")
res$bio <- c('a', 'a', 'a', 'a', 'a', 'a', 'a', 'b', 'b', 'b')

# Subsample the scone object
subsampled_res <- subsample_cells_with_min_bio(res, seed = 100, verbose = TRUE)


}
