% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sconeReport.R
\name{subsample_scone}
\alias{subsample_scone}
\title{Wrapper Function to subsample Scone object}
\usage{
subsample_scone(my_scone, subsample_gene_level = 100,
  subsample_cell_level = 100, cells_first = TRUE, at_bio = TRUE,
  keep_all_control = TRUE, seed = 100, verbose = FALSE)
}
\arguments{
\item{my_scone}{a \code{SconeExperiment} object}

\item{subsample_gene_level}{percent of genes you should take if subsampling genes by percent}

\item{subsample_cell_level}{percent of cells you should take if subsampling cells by percent}

\item{cells_first}{option to control wether to subsample cells and then genes, or the other way around}

\item{at_bio}{option to subsample to the number of cells of within the smallest bio group,
reducing every cell count in a bio group to the same number}

\item{keep_all_control}{option to keep all control genes and only
subsample remaining genes to fill to specificed percent,}

\item{seed}{the random seed}

\item{verbose}{option to display more information about subsampling}
}
\value{
The subsampled Scone Object
}
\description{
This function subsamples a scone object in a number of different ways
of a variety of normalization schemes.
}
\examples{
# Create a scone object
mat <- matrix(rpois(1000, lambda = 5), ncol=10)
colnames(mat) <- paste("X", 1:ncol(mat), sep="")
obj <- SconeExperiment(mat)
scone_object <- scone(obj, scaling=list(none=identity, uq=UQ_FN, deseq=DESEQ_FN),
           evaluate=TRUE, k_ruv=0, k_qc=0, eval_kclust=2,
           bpparam = BiocParallel::SerialParam())
qc = as.matrix(cbind(colSums(mat),colSums(mat > 0)))
rownames(qc) = colnames(mat)
colnames(qc) = c("NCOUNTS","NGENES")

scone_object$bio <- c('a', 'a', 'a', 'a', 'a', 'a', 'a', 'b', 'b', 'b')

# Subsample the scone object
subsampled_res <- subsample_scone(scone_object, subsample_gene_level = 50, subsample_cell_level = 100, 
                                     cells_first=TRUE, at_bio = TRUE,
                                     keep_all_control=TRUE, seed = 100, verbose = TRUE)


}
