% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCONE_DEFAULTS.R
\name{TMM_FN}
\alias{TMM_FN}
\title{TMM normalization wrapper.}
\usage{
TMM_FN(ei)
}
\arguments{
\item{ei}{= Numerical matrix. (rows = genes, cols = samples). Unique row.names are required.}
}
\value{
TMM normalized matrix (scaled by sample )
}
\description{
TMM normalization wrapper.
}

