% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCONE_DEFAULTS.R
\name{UQ_FN}
\alias{UQ_FN}
\title{Upper-quartile (UQ) scaling normalization wrapper function}
\usage{
UQ_FN(ei)
}
\arguments{
\item{ei}{Numerical matrix. (rows = genes, cols = samples).}
}
\value{
UQ normalized matrix.
}
\description{
Upper-quartile (UQ) scaling normalization wrapper function
}
\details{
SCONE scaling wrapper for \code{\link[edgeR]{calcNormFactors}}).
}
\examples{
ei <- matrix(0:20,nrow = 7)
eo <- UQ_FN(ei)

}
