% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCONE_DEFAULTS.R
\name{UQ_FN_POS}
\alias{UQ_FN_POS}
\title{Upper-quartile normalization applied to positive data.}
\usage{
UQ_FN_POS(ei)
}
\arguments{
\item{ei}{= Numerical matrix. (rows = genes, cols = samples). Unique row.names are required.}
}
\value{
Upper-quartile (positive) normalized matrix (scaled by sample UQ)
}
\description{
Upper-quartile normalization applied to positive data.
}

