% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCONE_DEFAULTS.R
\name{DESEQ_FN}
\alias{DESEQ_FN}
\title{DESeq size factor normalization wrapper.}
\usage{
DESEQ_FN(ei)
}
\arguments{
\item{ei}{Numerical matrix. (rows = genes, cols = samples).}
}
\value{
DESeq size factor normalized matrix.
}
\description{
DESeq size factor normalization wrapper.
}
\details{
SCONE scaling wrapper for
  \code{\link[DESeq]{estimateSizeFactorsForMatrix}}).
}
\examples{
ei <- matrix(0:20,nrow = 7)
eo <- DESEQ_FN(ei)

}

