% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCONE_DEFAULTS.R
\name{DESEQ_FN_POS}
\alias{DESEQ_FN_POS}
\title{DESeq normalization applied to positive data.}
\usage{
DESEQ_FN_POS(ei)
}
\arguments{
\item{ei}{= Numerical matrix. (rows = genes, cols = samples). Unique row.names are required.}
}
\value{
DESeq (positive) normalized matrix (scaled by sample )
}
\description{
DESeq normalization applied to positive data.
}

