% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCONE_DEFAULTS.R
\name{FQ_FN}
\alias{FQ_FN}
\title{Full-Quantile normalization wrapper.}
\usage{
FQ_FN(ei)
}
\arguments{
\item{ei}{= Numerical matrix. (rows = genes, cols = samples). Unique row.names are required.}
}
\value{
FQ normalized matrix.
}
\description{
Full-Quantile normalization wrapper.
}

