% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/helper.R
\docType{methods}
\name{get_params}
\alias{get_params}
\alias{get_params,SconeExperiment-method}
\title{Extract scone parameters}
\usage{
get_params(x)

\S4method{get_params}{SconeExperiment}(x)
}
\arguments{
\item{x}{an object of class \code{\link{SconeExperiment}}.}
}
\value{
A data.frame containing workflow parameters for each scone workflow.
}
\description{
Extract scone parameters
}
\examples{
set.seed(42)
mat <- matrix(rpois(500, lambda = 5), ncol=10)
colnames(mat) <- paste("X", 1:ncol(mat), sep="")
obj <- SconeExperiment(mat)
res <- scone(obj, scaling=list(none=identity, uq=UQ_FN),
           run = FALSE, k_ruv=0, k_qc=0, eval_kclust=2)
params = get_params(res)

}

