% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCONE_DEFAULTS.R
\name{SCRAN_FN}
\alias{SCRAN_FN}
\title{Pooled Sample normalization wrapper.}
\usage{
SCRAN_FN(ei)
}
\arguments{
\item{ei}{Numerical matrix. (rows = genes, cols = samples).}
}
\value{
SCRAN size factor normalized matrix.
}
\description{
SCONE scaling wrapper for \code{\link[scran]{computeSumFactors}}
  with clusters from \code{\link[scran]{quickCluster}}.
}
\examples{
set.seed(100)
popsize <- 800
ngenes <- 10000
all.facs <- 2^rnorm(popsize, sd=0.5)
mu1 = matrix(rep(all.facs*10,ngenes),nrow = ngenes,byrow = TRUE)
mu2 = matrix(c(rep(2^rnorm(ngenes, sd=0.5),popsize/2), rep(2^rnorm(ngenes, sd=0.5),popsize/2)),ncol = popsize)
ei <- matrix(rnbinom(ngenes*popsize, mu=as.vector(mu1*mu2), size=1), ncol=popsize)
eo <- SCRAN_FN(ei)

}

