% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCONE_DEFAULTS.R
\name{SCRAN_FN}
\alias{SCRAN_FN}
\title{Simple deconvolution normalization wrapper}
\usage{
SCRAN_FN(ei)
}
\arguments{
\item{ei}{Numerical matrix. (rows = genes, cols = samples).}
}
\value{
scran normalized matrix.
}
\description{
Simple deconvolution normalization wrapper
}
\details{
SCONE scaling wrapper for \code{\link[scran]{computeSumFactors}}).
}
\examples{
ei <- matrix(0:76,nrow = 7)
eo <- SCRAN_FN(ei)

}
