% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_filtering.R
\name{simple_FNR_params}
\alias{simple_FNR_params}
\title{Fit Logistic Regression Model of FNR against set of positive control (ubiquitously expressed) genes}
\usage{
simple_FNR_params(expr, pos_controls, fn_tresh = 0.01)
}
\arguments{
\item{expr}{matrix The data matrix in transcript-proportional units (genes in rows, cells in columns).}

\item{pos_controls}{A boolean vector indexing positive control genes that will be used to compute false-negative rate characteristics.}

\item{fn_tresh}{Inclusive threshold for negative detection. Default 0.01.}
}
\value{
A list of logistic regression coefficients corresponding to glm fits in each sample. If a fit did not converge, the result reported is NA.
}
\description{
Fit Logistic Regression Model of FNR against set of positive control (ubiquitously expressed) genes
}
\details{
logit(Probability of False Negative) ~ a + b*(mean log10p1 expression) .
}

