% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCONE_DEFAULTS.R
\name{TMM_FN}
\alias{TMM_FN}
\title{Weighted trimmed mean of M-values (TMM) normalization wrapper.}
\usage{
TMM_FN(ei)
}
\arguments{
\item{ei}{Numerical matrix. (rows = genes, cols = samples).}
}
\value{
TMM normalized matrix.
}
\description{
Weighted trimmed mean of M-values (TMM) normalization wrapper.
}
\details{
SCONE scaling wrapper for \code{\link[edgeR]{calcNormFactors}}).
}
\examples{
ei <- matrix(0:20,nrow = 7)
eo <- TMM_FN(ei)

}

