% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCONE_DEFAULTS.R
\name{UQ_FN}
\alias{UQ_FN}
\title{Upper-quartile normalization wrapper.}
\usage{
UQ_FN(ei)
}
\arguments{
\item{ei}{= Numerical matrix. (rows = genes, cols = samples). Unique row.names are required.}
}
\value{
Upper-quartile normalized matrix (scaled by sample UQ)
}
\description{
Upper-quartile normalization wrapper.
}

