% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCONE_DEFAULTS.R
\name{UQ_FN_POS}
\alias{UQ_FN_POS}
\title{Upper-quartile normalization derived from positive data.}
\usage{
UQ_FN_POS(ei)
}
\arguments{
\item{ei}{Numerical matrix. (rows = genes, cols = samples).}
}
\value{
Upper-quartile (positive) normalized matrix.
}
\description{
Upper-quartile normalization derived from positive data.
}
\details{
SCONE scaling function scales expression data by upper quartile of
  positive data.
}
\examples{
ei <- matrix(0:20,nrow = 7)
ei[1:3,] <- 0
eo <- UQ_FN_POS(ei)

}

